agroai_api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
httpx==0.26.0
prometheus-client==0.19.0
python-json-logger==2.0.7
tenacity==8.2.3

# Cloud Storage
boto3==1.34.13
azure-storage-blob==12.19.0
watchdog==3.0.0

# ML & Data Processing
scikit-learn==1.4.0
pandas==2.1.4
numpy==1.26.3
joblib==1.3.2

# Rate Limiting
slowapi==0.1.9

# CLI
click==8.1.7
rich==13.7.0

# Testing
pytest==7.4.4
pytest-mock==3.12.0
moto==4.2.14
faker==22.0.0
pytest-asyncio==0.23.3
pytest-cov==4.1.0
ruff==0.1.14
mypy==1.8.0
types-passlib==1.7.7.20240106
types-python-jose==3.3.4.20240106
black==24.1.1
